Refactor functions in modal.js
